requirements.txt
flask
opencv-python
requests
python-dotenv
pillow